feat: Show current slide
